pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-codeforge-pro-env"
version = "0.1.0"
description = "CodeForge Pro – Real-world dev agent environment for OpenEnv (code review, bug triage, pipeline debugging)"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
server = "codeforge_pro_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["codeforge_pro_env", "codeforge_pro_env.server"]
package-dir = { "codeforge_pro_env" = ".", "codeforge_pro_env.server" = "server" }